+++ kernel/src/globalesKernel.h
@@ -119,7 +119,7 @@ extern int tiempo_suspension;
 extern char* puerto_escucha_dispatch;
 extern char* puerto_escucha_interrupt;
 extern char* puerto_escucha_IO;
-extern int alfa;
+extern double alfa;
 extern int64_t estimacion_inicial;
 extern char*  algoritmo_cola_new;
